+++ bot/commands/fun/erase.js
@@ -1,9 +1,36 @@
 'use strict';
-/*
-ONLY MODS
-!erase <trigger name>
-*/
+var repo = require(process.cwd()+'/repo');
 
-module.exports = function(bot, db) {
-  bot.sendChat('The ability to erase a trigger is coming soon!');
+function displayHelp(bot){
+  bot.sendChat('*usage:* !erase <trigger_name>');
+  bot.sendChat('Don\'t add the "!" when erasing a trigger');
+}
+
+module.exports = function(bot, db, data) {
+  if (data.params === void(0) || data.params.length < 1) {
+    return displayHelp(bot);
+  }
+
+  // if not a MOD, GTFO!
+  if ( !bot.hasPermission(bot.getUserByName(data.user.username), 'skip') ) {
+    return bot.sendChat('Ah ah ah, not without the magic word! (only mods can do this)');
+  }
+
+  if (data.params[0].charAt(0) === '!') {
+    displayHelp(bot);
+    return;
+  }
+
+  data.triggerName = data.params[0];
+
+  repo.getTrigger(bot, db, data.triggerName, function(val){
+
+    if (val !== null && data.params.length === 1) {
+      // deleting a trigger
+      bot.sendChat('Trigger erase functionality is coming soon');
+      return;
+    }
+
+  });
+
 };